refactor: extract facet_distribution tests to separate file
Move 7 test functions from facet_distribution.rs inline test module
to a dedicated facet_distribution_test.rs file, reducing
facet_distribution.rs from 889 to 437 lines (51% reduction).
All tests pass unchanged.

Part of #5116
